Server down handling and server volume control

+++ src/com/mindplus/model/ClientListController.java
@@ -40,7 +40,7 @@ public class ClientListController {
 		}
 	}
 
-	public synchronized int size() {
+	public synchronized long size() {
 		return clientList.size();
 	}
 }
